requirement.txt
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.2
SQLAlchemy==1.4.44
mysql-connector-python==8.0.29
alembic==1.9.1
click==8.1.3
dotenv-python==0.0.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Migrate==4.0.0
Flask-SQLAlchemy
greenlet==2.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
mysql-connector-python==8.0.29
protobuf==4.21.12
python-dotenv==0.21.0
six==1.16.0
SQLAlchemy==1.4.44
Werkzeug==2.2.2
flask_sqlalchemy
